SHARK4R: Retrieving, Analyzing, and Validating Marine Data from SHARK and Nordic Microalgae

R package for data analysis, automatic quality control and data retrieval for physical, chemical and marine biological data generated by SHARK and Nordic Microalgae, developed by SMHI. Parts have been modified from `obistools`. Provoost P and Bosch S (2018). "obistools: Tools for data enhancement and quality control." Ocean Biogeographic Information System. Intergovernmental Oceanographic Commission of UNESCO. https://iobis.github.io/obistools/.

See also

Author

Maintainer: Anders Torstensson anders.torstensson@smhi.se (ORCID) (Swedish Meteorological and Hydrological Institute)

Authors:

Markus Lindh (ORCID) (Swedish Meteorological and Hydrological Institute)

Other contributors:

SHARK shark@smhi.se [copyright holder]