Construct Dyntaxa taxonomy table from API

This function constructs a taxonomy table based on Dyntaxa parent taxon IDs. It queries the SLU Artdatabanken API (Dyntaxa) to fetch taxonomy information and organizes the data into a hierarchical table.

Usage

construct_dyntaxa_table(
  parent_ids,
  subscription_key,
  shark_output = TRUE,
  recommended_only = TRUE,
  add_genus_children = FALSE,
  verbose = TRUE
)

Arguments

parent_ids

A list containing parent taxon IDs for which taxonomy information is requested.

subscription_key

A character string containing the subscription key for accessing the SLU Artdatabanken API. A key is provided for registered users at Artdatabanken.

shark_output

Logical. If TRUE, the function will return selected column headers that match SHARK output. If FALSE, all columns are returned. Default is TRUE.

recommended_only

Logical. If TRUE, the function will return only recommended (accepted) names. If FALSE, all names are returned. Default is TRUE.

add_genus_children

Logical. If TRUE, the output will include children from all valid genera taxon_ids present in parent_ids. Default is FALSE.

verbose

Logical. Default is TRUE.

Value

A data frame representing the constructed taxonomy table.

Details

Please read the conditions and register before using the API.

Data collected through the API is stored at SLU Artdatabanken.

Genom att anvanda denna applikation atar jag mig att folja regler for anvandning av SLU Artdatabankens information, inklusive att respektera tredje mans upphovsratt. Detta atagande gors aven direkt mot SLU. Vid brott mot detta kan jag forlora ratten att anvanda applikationen

See also

Examples

if (FALSE) { # \dontrun{
# Construct Dyntaxa taxonomy table for taxon IDs 238366 and 1010380
parent_ids <- get_dyntaxa_parent_ids(c(238366, 1010380), "your_subscription_key")
taxonomy_table <- construct_dyntaxa_table(parent_ids, "your_subscription_key")
print(taxonomy_table)
} # }