Get geographical information from coordinates

This function was deprecated because SHARKdata has now been replaced by the SHARK API. Use get_shark_table() instead.

This function enriches the provided data with geographical information based on latitude and longitude coordinates in the Baltic Sea, Kattegat and Skagerrak.

Usage

get_geographical_info(latitude_dd, longitude_dd)

Arguments

latitude_dd

A numeric vector representing latitude coordinates in decimal degrees.

longitude_dd

A numeric vector representing longitude coordinates in decimal degrees.

Value

A data frame containing the original coordinates with the added geographical information location_type_area, location_svar_sea_area_name and location_sea_basin.

Details

The function reads shapefiles and basin names to assign geographical information to the dataset. It provides details about the location type area and the Svar Sea area name based on the given coordinates.

See also

Examples

if (FALSE) { # \dontrun{
latitude <- c(58.25830, 58.59367, 63.49983)
longitude <- c(11.43330, 18.23550, 19.81933)
dataset_with_geo_info <- get_geographical_info(latitude, longitude)
} # }