Retrieve media metadata from Nordic Microalgae

This function retrieves metadata for media items from the Nordic Microalgae API. It returns detailed attributes such as title, caption, location, sampling date, geographic coordinates, and contributor information.

Usage

get_nua_media_metadata(unparsed = FALSE)

Arguments

unparsed

Logical. If TRUE, complete API response is returned as an unparsed list. Default is FALSE.

Value

When unparsed = FALSE: a tibble with the following columns:

slug: The slug of the media item.

taxon_slug: The slug of the related taxon.

scientific_name: The scientific name of the related taxon.

file: The filename of the media item.

type: The MIME type of the media item.

title: The title of the media item.

caption: The caption of the media item.

license: The license of the media item.

location: The location where the media was collected.

contributor: The contributor of the media item.

photographer_artist: The photographer or artist.

copyright_holder: The copyright holder.

galleries: Comma-separated list of galleries.

technique: The imaging technique used.

contrast_enhancement: The contrast enhancement method used.

preservation: The preservation method used.

stain: The stain used.

sampling_date: The date the sample was collected.

geographic_area: The geographic area of collection.

latitude_degree: The latitude in degrees.

longitude_degree: The longitude in degrees.

institute: Comma-separated list of institutes.

contributing_organisation: The contributing organisation.

created_at: The creation timestamp.

updated_at: The last update timestamp.

See also

Examples

# \donttest{
# Retrieve media metadata
media_metadata <- get_nua_media_metadata(unparsed = FALSE)

# Preview the extracted data
head(media_metadata)
#> # A tibble: 6 × 25
#>   slug     taxon_slug scientific_name file  type  title caption license location
#>   <chr>    <chr>      <chr>           <chr> <chr> <chr> <chr>   <chr>   <chr>   
#> 1 octacti… octactis-… Octactis specu… octa… imag… Octa… "Two c… Creati… Danafjo…
#> 2 lennoxi… lennoxia-… Lennoxia faveo… lenn… imag… Lenn… ""      Creati… Danafjo…
#> 3 bolmen-… NA         NA              bolm… imag… Bolm… "Taken… Creati… Bolmen,…
#> 4 bolmen-… NA         NA              bolm… imag… Bolm… "Taken… Creati… Bolmen,…
#> 5 bolmen-… NA         NA              bolm… imag… Bolm… "Taken… Creati… Bolmen,…
#> 6 hasslem… NA         NA              hass… imag… Hass… "Taken… Creati… Hasslem…
#> # ℹ 16 more variables: contributor <chr>, photographer_artist <chr>,
#> #   copyright_holder <chr>, galleries <chr>, technique <list>,
#> #   contrast_enhancement <list>, preservation <list>, stain <list>,
#> #   sampling_date <chr>, geographic_area <chr>, latitude_degree <chr>,
#> #   longitude_degree <chr>, institute <chr>, contributing_organisation <chr>,
#> #   created_at <chr>, updated_at <chr>
# }