Retrieve Data from SHARK API

The get_shark_data function retrieves data from the SHARK database hosted by SMHI. The function sends a POST request to the SHARK API with customizable filters, including year, month, taxon name, water category, and more, and returns the retrieved data as a structured data.frame. To view available filter options, see get_shark_options.

Usage

get_shark_data(
  tableView = "sharkweb_overview",
  headerLang = "internal_key",
  hideEmptyColumns = FALSE,
  fromYear = 2019,
  toYear = 2020,
  months = c(),
  dataTypes = c(),
  parameters = c(),
  checkStatus = "",
  qualityFlags = c(),
  deliverers = c(),
  orderers = c(),
  projects = c(),
  datasets = c(),
  minSamplingDepth = "",
  maxSamplingDepth = "",
  redListedCategory = c(),
  taxonName = c(),
  stationName = c(),
  vattenDistrikt = c(),
  seaBasins = c(),
  counties = c(),
  municipalities = c(),
  waterCategories = c(),
  typOmraden = c(),
  helcomOspar = c(),
  seaAreas = c(),
  prod = TRUE,
  save_data = FALSE,
  file_path = NULL,
  verbose = TRUE
)

Arguments

tableView

Character. Specifies the view of the table to retrieve. Options include:

"sharkweb_overview"

"sharkweb_all"

"sharkdata_bacterioplankton"

"sharkdata_chlorophyll"

"sharkdata_epibenthos"

"sharkdata_greyseal"

"sharkdata_harbourporpoise"

"sharkdata_harbourseal"

"sharkdata_jellyfish"

"sharkdata_physicalchemical_columns"

"sharkdata_phytoplankton"

"sharkdata_picoplankton"

"sharkdata_planktonbarcoding"

"sharkdata_primaryproduction"

"sharkdata_ringedseal"

"sharkdata_sealpathology"

"sharkdata_sedimentation"

"sharkdata_zoobenthos"

"sharkdata_zooplankton"

"report_sum_year_param"

"report_sum_year_param_taxon"

"report_sampling_per_station"

"report_obs_taxon"

"report_stations"

"report_taxon"

Default is "sharkweb_overview".

headerLang

Character. Language option for column headers. Possible values:

"sv": Swedish.

"en": English.

"short": Shortened version.

"internal_key": Internal key (default).

hideEmptyColumns

Logical. Whether to hide empty columns. Default is FALSE.

fromYear

Integer. Starting year for data retrieval. Default is 2019.

toYear

Integer. Ending year for data retrieval. Default is 2020.

months

Integer vector. The months to retrieve data for, e.g., c(4, 5, 6) for April to June.

dataTypes

Character vector. Specifies data types to filter, such as "Chlorophyll", "Epibenthos", etc.

parameters

Character vector. Optional parameters to filter the results by, such as "Chlorophyll-a".

checkStatus

Character string. Optional status check to filter results.

qualityFlags

Character vector. Quality flags to filter the data.

deliverers

Character vector. Specifies the data deliverers to filter by.

orderers

Character vector. Orderers to filter by specific organizations or individuals.

projects

Character vector. Projects to filter data by specific research or monitoring projects.

datasets

Character vector. Datasets to filter data by specific datasets.

minSamplingDepth

Numeric. Minimum sampling depth (in meters) to filter the data.

maxSamplingDepth

Numeric. Maximum sampling depth (in meters) to filter the data.

redListedCategory

Character vector. Red-listed taxa for conservation filtering.

taxonName

Character vector. Optional vector of taxa names to filter by.

stationName

Character vector. Station names to filter data by specific stations.

vattenDistrikt

Character vector. Water district names to filter by Swedish water districts.

seaBasins

Character vector. Sea basins to filter by.

counties

Character vector. Counties to filter by specific administrative regions.

municipalities

Character vector. Municipalities to filter by.

waterCategories

Character vector. Water categories to filter by.

typOmraden

Character vector. Type areas to filter by.

helcomOspar

Character vector. HELCOM or OSPAR areas for regional filtering.

seaAreas

Character vector. Sea area codes to filter by specific sea areas.

prod

Logical. Whether to query the PROD (production) server or the TEST (testing) server. Default is TRUE (PROD).

save_data

Logical. If TRUE, the data will be saved to a specified file in UTF-8 encoding. If FALSE, a temporary file will be created instead.

file_path

Character. The file path where the data should be saved. Required if save_data is TRUE. Ignored if save_data is FALSE.

verbose

Logical. Whether to display progress information. Default is TRUE.

Value

A data.frame containing the retrieved SHARK data, with column names based on the API's response.

Details

This function sends a POST request to the SHARK API with the specified filters. The response is parsed as JSON and then converted into a data.frame. The function handles the dynamic construction of the query body to filter the data based on the provided parameters.

See also

Examples

if (FALSE) { # \dontrun{
  # Retrieve chlorophyll data from 2019 to 2020 for April to June
  shark_data <- get_shark_data(fromYear = 2019, toYear = 2020,
                               months = c(4, 5, 6), dataTypes = c("Chlorophyll"))
  View(shark_data)
} # }