Load dataset names from SHARKdata based on type, year, and data deliverer

This function was deprecated because SHARKdata has now been replaced by the SHARK API. Use get_shark_table() instead.

This function gets dataset names fron the SHARKdata API based on the specified data type, year, and data deliverer.

Usage

load_dataset_names(dataset_type, year = NA, data_deliverer = NA)

Arguments

dataset_type

The type of dataset to load.

year

The year to filter datasets based on dataset name (default NA results in all available years). Please note that dataset names can contain multiple years (e.g. a dataset with data from 2012-2014 will not be selected when specifying year 2013 in argument). Consider filter data after download for more extensive data.

data_deliverer

The data deliverer (code) to filter datasets (default is NA, results in all deliverers). See SMHI codelist for available codes. Consider filter data after download for more extensive data.

Value

A data frame containing filtered dataset names.

See also

Examples

if (FALSE) { # \dontrun{
# All Zoobenthos datasets
load_dataset_names("Zoobenthos")

# All Phytoplankton datasets from 2012:2020 from SMHI and UMSC.
load_dataset_names("Phytoplankton", year = 2012:2020, data_deliverer = c("SMHI", "UMSC"))
} # }