Retrieve and Organize WoRMS Taxonomy for SHARK Aphia IDs

This function was deprecated and replaced by a more accurate name. Use add_worms_taxonomy() instead.

This function collects WoRMS (World Register of Marine Species) taxonomy information for a given set of Aphia IDs. The data is organized into a full taxonomic table that can be joined with data downloaded from SHARK.

Usage

update_worms_taxonomy(aphia_id, aphiaid = deprecated())

Arguments

aphia_id

A numeric vector containing Aphia IDs for which WoRMS taxonomy needs to be updated.

aphiaid

Use aphia_id instead.

Value

A data frame containing updated WoRMS taxonomy information.

See also

Examples

if (FALSE) { # \dontrun{
# Update WoRMS taxonomy for a set of Aphia IDs
updated_taxonomy <- update_worms_taxonomy(c(149619, 149122, 11))
print(updated_taxonomy)
} # }